pyproject.toml
[project]
name = "streamlit-user-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.6",  # Updated from 1.3.2 to latest 1.6.6
    "httpx>=0.28.1",
    "msal>=1.34.0",  # Already at latest version
    "numpy>=2.4.1",
    "pandas<2.4.0",
    "plotly>=6.5.2",
    "python-jose[cryptography]>=3.5.0",  # Already at latest version
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.46",
    "streamlit-nightly>=1.53.2.dev20260128", # Latest Jan 2026 Nightly
    "uvloop>=0.21.0", # Performance boost for Streamlit
    "orjson>=3.11.5",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.29.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
